+++ src/abcd_graph/utils.py
@@ -21,19 +21,13 @@
 
 import random
 from functools import wraps
-from typing import (
-    TYPE_CHECKING,
-    Callable,
-)
+from typing import Callable
 
 import numpy
 from typing_extensions import (
     ParamSpec,
     TypeVar,
 )
-
-if TYPE_CHECKING:
-    from abcd_graph.abcd import Graph
 
 P = ParamSpec("P")
 R = TypeVar("R")
@@ -59,14 +53,3 @@ def require(package_name: str) -> Callable[[Callable[P, R]], Callable[P, R]]:
         return wrapper
 
     return deco
-
-
-def require_graph_built(func: Callable[P, R]) -> Callable[P, R]:
-    @wraps(func)
-    def wrapper(self: "Graph", *args: P.args, **kwargs: P.kwargs) -> R:
-        if not self.is_built:
-            raise RuntimeError("Graph has not been built yet")
-
-        return func(self, *args, **kwargs)  # type: ignore[arg-type]
-
-    return wrapper  # type: ignore[return-value]