+++ CHANGELOG.md
@@ -7,3 +7,9 @@
 - Add exporting graph to vector of pairs ([#19](https://github.com/AleksanderWWW/abcd-graph/pull/19))
 - Added empirical xi (global and per community) to diagnostics ([#20](https://github.com/AleksanderWWW/abcd-graph/pull/20))
 - Added `degree_sequence` property calculated from edge list ([#20](https://github.com/AleksanderWWW/abcd-graph/pull/20))
+- Implemented callback mechanism to handle diagnostics and visualization ([#23](https://github.com/AleksanderWWW/abcd-graph/pull/23))
+- Added visualization of CDF of community sizes and degrees ([#23](https://github.com/AleksanderWWW/abcd-graph/pull/23))
+
+### Changes
+- Restructured the project to have a more consistent API ([#23](https://github.com/AleksanderWWW/abcd-graph/pull/23))
+- Renamed env variable controlling logging level to `ABCD_LOG` ([#23](https://github.com/AleksanderWWW/abcd-graph/pull/23))